+++ src/Controller/TrabajadoresController.php
@@ -25,6 +25,10 @@ class TrabajadoresController extends AppController
      */
     public function index()
     {
+        if($this->request->session()->read('Auth.User.funcion')=='Visitante') {
+            $this->Flash->error(__('Usted no tiene permiso para acceder a la pagina solicitada.'));
+            return $this->redirect($this->referer());
+        }
         $trabajadores = $this->paginate($this->Trabajadores);
 
         $this->set(compact('trabajadores'));
@@ -33,6 +37,7 @@ class TrabajadoresController extends AppController
 
     public function menu()
     {
+
         $trabajadores = $this->paginate($this->Trabajadores);
 
         $this->set(compact('trabajadores'));
@@ -47,6 +52,13 @@ class TrabajadoresController extends AppController
      */
     public function view($id = null)
     {
+        if($this->request->session()->read('Auth.User.funcion')=='Visitante') {
+            if($this->request->session()->read('Auth.User.trabajador_id')!=$id)
+            {
+                $this->Flash->error(__('Usted no tiene permiso para acceder a la pagina solicitada.'));
+                return $this->redirect($this->referer());
+            }
+        }
         $trabajador = $this->Trabajadores->get($id, [
             'contain' => ['Procesos', 'Contratos', 'Usuarios']
         ]);
@@ -62,6 +74,10 @@ class TrabajadoresController extends AppController
      */
     public function add()
     {
+        if($this->request->session()->read('Auth.User.funcion')=='Visitante'||$this->request->session()->read('Auth.User.funcion')=='Operador') {
+            $this->Flash->error(__('Usted no tiene permiso para acceder a la pagina solicitada.'));
+            return $this->redirect($this->referer());
+        }
         $trabajador = $this->Trabajadores->newEntity();
         if ($this->request->is('post')) {
             $trabajador = $this->Trabajadores->patchEntity($trabajador, $this->request->data);
@@ -115,8 +131,6 @@ class TrabajadoresController extends AppController
             $trabajador = $this->Trabajadores->patchEntity($trabajador, $this->request->data);
             if ($this->Trabajadores->save($trabajador)) {
                 $this->Flash->success(__('El trabajador ha sido registrado.'));
-
-
                 $usuario = TableRegistry::get('Usuarios')->newEntity();
                 $username=$this->request->data['apellido'].$this->request->data['nombre'][0];
                 $username=$this->getnewname($username);
@@ -151,6 +165,13 @@ class TrabajadoresController extends AppController
      */
     public function edit($id = null)
     {
+        if($this->request->session()->read('Auth.User.funcion')=='Visitante'||$this->request->session()->read('Auth.User.funcion')=='Operador') {
+            if($this->request->session()->read('Auth.User.trabajador_id')!=$id)
+            {
+                $this->Flash->error(__('Usted no tiene permiso para acceder a la pagina solicitada.'));
+                return $this->redirect($this->referer());
+            }
+        }
         $trabajador = $this->Trabajadores->get($id, [
             'contain' => ['Procesos']
         ]);
@@ -178,6 +199,10 @@ class TrabajadoresController extends AppController
      */
     public function delete($id = null)
     {
+        if($this->request->session()->read('Auth.User.funcion')!='Superadministrador') {
+            $this->Flash->error(__('Usted no tiene permiso para acceder a la accion solicitada.'));
+            return $this->redirect($this->referer());
+        }
         $this->request->allowMethod(['post', 'delete']);
         $trabajador = $this->Trabajadores->get($id);
         if ($this->Trabajadores->delete($trabajador)) {
